tests: add Go client suite + fix wire-format type-name compat
The Go bigquery client (cloud.google.com/go/bigquery) parses query
result values via a switch on the legacy type-name constants only —
INTEGER, FLOAT, BOOLEAN, RECORD — with no fallback through the
fieldAliases map for the standard SQL names INT64 / FLOAT64 / BOOL /
STRUCT. Emitting the standard names broke iteration with
"unrecognized type: INT64". Real BigQuery's REST API responses use
the legacy names too; Python + Node clients accept both.

Adds bqTypeToWire() that maps internal BqType → legacy wire name and
plumbs it through fieldToWire() in both src/routes/tables.ts and
src/sql/queryEngine.ts. JS tests that asserted on standard names are
updated to expect the legacy form.

New Go client test suite at test/clients/go/ covers basic CRUD, SQL
features (parameters, MERGE/QUALIFY/PIVOT, CTE+window, DML, scripts,
transactions, dry-run, wildcard tables), partitioning, routines +
views + MVs, INFORMATION_SCHEMA, load/extract/copy jobs, and
advanced features (labels, locations, query cache, multi-project
isolation). Spawns the emulator subprocess + an in-process GCS stub
from TestMain, scoped per-test by unique project id.

CI: new go-client matrix job (Go 1.25 + 1.26) running the suite.

+++ test/api/dry-run.test.ts
@@ -146,9 +146,9 @@ test('POST /queries dryRun=true returns schema with no rows', async () => {
   // Schema is filled in.
   assert.equal(body.schema.fields.length, 2);
   assert.equal(body.schema.fields[0]?.name, 'one');
-  assert.equal(body.schema.fields[0]?.type, 'INT64');
+  assert.equal(body.schema.fields[0]?.type, 'INTEGER');
   assert.equal(body.schema.fields[1]?.name, 'two');
-  assert.equal(body.schema.fields[1]?.type, 'FLOAT64');
+  assert.equal(body.schema.fields[1]?.type, 'FLOAT');
 });
 
 test('POST /queries dryRun against a real table types the projection correctly', async () => {
@@ -166,7 +166,7 @@ test('POST /queries dryRun against a real table types the projection correctly',
   );
   assert.deepEqual(
     body.schema.fields.map((f) => f.type),
-    ['STRING', 'INT64'],
+    ['STRING', 'INTEGER'],
   );
 });
 
@@ -248,7 +248,7 @@ test('POST /jobs dryRun with parameters validates the parameter shape', async ()
   assert.equal(status, 200);
   const body = json as JobResourceWire;
   assert.equal(body.statistics.query?.schema?.fields[0]?.name, 'doubled');
-  assert.equal(body.statistics.query?.schema?.fields[0]?.type, 'INT64');
+  assert.equal(body.statistics.query?.schema?.fields[0]?.type, 'INTEGER');
 });
 
 test('POST /jobs dryRun: invalid SQL → 400', async () => {
